+++ app/src/androidTest/java/com/dheeraj/news/data/repository/FakeRepository.kt
@@ -1,10 +1,10 @@
 package com.dheeraj.news.data.repository
 
-import com.dheeraj.news.data.model.CommentsResponse
-import com.dheeraj.news.data.model.LikesResponse
-import com.dheeraj.news.domain.entity.NewsArticle
+import com.dheeraj.news.data.entity.CommentsResponse
+import com.dheeraj.news.data.entity.LikesResponse
+import com.dheeraj.news.domain.model.NewsArticle
 import com.dheeraj.news.domain.repository.NewsRepository
-import com.dheeraj.news.util.Resource
+import com.dheeraj.news.data.util.Resource
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
