+++ app/src/main/java/com/dheeraj/news/data/mappers/NewsArticleMapper.kt
@@ -1,17 +1,17 @@
 package com.dheeraj.news.data.mappers
 
-import com.dheeraj.news.data.model.Article
-import com.dheeraj.news.domain.entity.NewsArticle
-import com.dheeraj.news.util.EntityMapper
+import com.dheeraj.news.data.entity.Article
+import com.dheeraj.news.domain.model.NewsArticle
+import com.dheeraj.news.data.util.EntityMapper
 
 class NewsArticleMapper : EntityMapper<NewsArticle, Article> {
 
-    override fun mapToEntity(model: Article): NewsArticle {
+    override fun mapFromEntity(entity: Article): NewsArticle {
         return NewsArticle(
-            articleId = model.url,
-            description = model.description,
-            author = model.author,
-            imageUrl = model.urlToImage ?: NewsArticle.PLACEHOLDER_IMAGE_URL
+            articleId = entity.url,
+            description = entity.description,
+            author = entity.author,
+            imageUrl = entity.urlToImage ?: NewsArticle.PLACEHOLDER_IMAGE_URL
         )
     }
 }
