+++ app/src/main/java/com/dheeraj/news/data/repository/dataSource/NewsRemoteDataSource.kt
@@ -1,9 +1,9 @@
 package com.dheeraj.news.data.repository.dataSource
 
-import com.dheeraj.news.data.model.CommentsResponse
-import com.dheeraj.news.data.model.LikesResponse
-import com.dheeraj.news.data.model.NewsHeadlinesApiResponse
-import com.dheeraj.news.util.Resource
+import com.dheeraj.news.data.entity.CommentsResponse
+import com.dheeraj.news.data.entity.LikesResponse
+import com.dheeraj.news.data.entity.NewsHeadlinesApiResponse
+import com.dheeraj.news.data.util.Resource
 
 interface NewsRemoteDataSource {
