+++ app/src/main/java/com/dheeraj/news/data/util/EntityMapper.kt
@@ -1,0 +1,6 @@
+package com.dheeraj.news.data.util
+
+interface EntityMapper<Model, Entity> {
+
+    fun mapFromEntity(entity: Entity) : Model
+}
