+++ app/src/main/java/com/dheeraj/news/data/util/SafeApiCall.kt
@@ -1,4 +1,4 @@
-package com.dheeraj.news.util
+package com.dheeraj.news.data.util
 
 import retrofit2.Response
