+++ app/src/main/java/com/dheeraj/news/domain/repository/NewsRepository.kt
@@ -1,9 +1,9 @@
 package com.dheeraj.news.domain.repository
 
-import com.dheeraj.news.data.model.CommentsResponse
-import com.dheeraj.news.data.model.LikesResponse
-import com.dheeraj.news.domain.entity.NewsArticle
-import com.dheeraj.news.util.Resource
+import com.dheeraj.news.data.entity.CommentsResponse
+import com.dheeraj.news.data.entity.LikesResponse
+import com.dheeraj.news.domain.model.NewsArticle
+import com.dheeraj.news.data.util.Resource
 import kotlinx.coroutines.flow.Flow
 
 interface NewsRepository {
