--- app/src/main/java/com/dheeraj/news/util/EntityMapper.kt
@@ -1,6 +1,0 @@
-package com.dheeraj.news.util
-
-interface EntityMapper<Entity, Model> {
-
-    fun mapToEntity(model: Model) : Entity
-}
