+++ app/src/main/res/layout/fragment_news_article.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="newsArticle"
-            type="com.dheeraj.news.domain.entity.NewsArticle" />
+            type="com.dheeraj.news.domain.model.NewsArticle" />
 
         <variable
             name="view"
