+++ app/src/test/java/com/dheeraj/news/dataSourceImpl/NewsRemoteDataSourceImplTest.kt
@@ -3,7 +3,7 @@ package com.dheeraj.news.dataSourceImpl
 import com.dheeraj.news.data.api.NewsApiService
 import com.dheeraj.news.data.repository.dataSource.NewsRemoteDataSource
 import com.dheeraj.news.data.repository.dataSourceImpl.NewsRemoteDataSourceImpl
-import com.dheeraj.news.util.Resource
+import com.dheeraj.news.data.util.Resource
 import com.google.common.truth.Truth.*
 import kotlinx.coroutines.runBlocking
 import okhttp3.mockwebserver.MockResponse
